* commit '191f131b123e20a217bcc57a3b2b626cc39f6bac':
  Fix build break for size_t vs. uint32_t difference.

+++ cpu_ref/rsCpuIntrinsic.h
@@ -39,7 +39,7 @@ public:
 
     virtual void invokeForEachMulti(uint32_t slot,
                        const Allocation ** ain,
-                       size_t inLen,
+                       uint32_t inLen,
                        Allocation * aout,
                        const void * usr,
                        uint32_t usrLen,
